requirements.txt
requests>=2.31.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
